fix: refactor healthcheck controller
