+++ app/main.go
@@ -5,5 +5,5 @@ import (
 )
 
 func main() {
-	cmd.StartRest()
+	cmd.InitRest()
 }
